Simplifying the communication with extension
