+++ src/utils/extension/browserApi.js
@@ -1,0 +1,6 @@
+/*global chrome*/
+/*global browser*/
+
+import { runningInChromeDesktop } from "../misc/browserDetection";
+
+export const BROWSER_API = runningInChromeDesktop() ? chrome : browser;
